Refatoring claim to its own file, no funcitonal change.
