+++ src/day3/CMakeLists.txt
@@ -1,5 +1,5 @@
 
 new_cpp_executeable(
         NAME day3
-        SOURCES main.cpp manager.cpp manager.hpp
+        SOURCES main.cpp manager.cpp manager.hpp claim.hpp claim.cpp
         )
